lots of cool new features, dsl and refactoring - updated README to reflect changes
