+++ README.md
@@ -3,26 +3,46 @@
 A gem/plugin to generate those new social buttons like *tweet*, *like*, *pinit*
 
 Currently this gem supports 3 social buttons:
+
 * Twitter *tweet* button
 * Facebook *like* button
 * Pinterest *pinit* button
 
+The gem is a Rails 3+ engine!
 
 # Installation
+
 Add the gem to your Gemfile and use Bundler to install the gem.
 
 ```ruby
-gem "social-buttons", git: "git://github.com/iffyuva/social-buttons.git"
+gem "social-buttons", git: "git://github.com/kristianmandrup/social-buttons.git"
 ```
 
 # Usage
 
-First, include `TweetButton` or `LikeButton` or `PinitButton` into your application helper. After that, using it is as simple as adding a single method call to your views:
+This gem is a Rails 3+ compatible engine and now automatically includes all the view helper modules into `ActionView::Base`
 
+This gem includes the following view helper modules in `social_buttons/view_helpers`:
+
+* `Tweet`
+* `Like`
+* `Pinit` 
+
+Using SocialButtons is as simple as adding a single method call to your views:
+
+ERB example:
 ```erb
 <%= tweet_button %>
 <%= like_button(app_id) %>
 <%= pinit_button %>
+```
+
+HAML example:
+
+```haml
+= tweet_button
+= like_button(app_id)
+= pinit_button
 ```
 
 Bam.  Done.  You'll have a sweet lookin' Social Buttons all up in your view.
@@ -41,11 +61,13 @@ Of course, you can customize it. Depending on the type of button, there are diff
 More information can be found from [Twitter](https://twitter.com/about/resources/buttons#tweet)
 
 So, if you wanted to tweet about Hacker News, attribute it to Peter Cooper, and add some custom text, all from a tweet button with a horizontal counter, you'd do this:
+
 ```erb
 <%= tweet_button(via: => "peterc", url: "http://news.ycombinator.com", :text => "AWESOME.") %>
 ```
 
 ### Like Button
+
 NOTE: Facebook App Id is mandatory
 
 * `:href` - The URL to like, the default is the current URL.
@@ -60,6 +82,7 @@ NOTE: Facebook App Id is mandatory
 More information can be found from [Facebook](http://developers.facebook.com/docs/reference/plugins/like/)
 
 So, a simple like button can be added like this:
+
 ```erb
 <%= like_button(app_id) %>
 ```
@@ -74,30 +97,60 @@ So, a simple like button can be added like this:
 More information can be found from [Pinterest](http://pinterest.com/about/goodies/)
 
 So, a simple pinit button can be added like this:
+
 ```erb
 <%= pinit_button(url: request.url, media: @picture.image.url, description: "Amazing Picture") %>
 ```
 
 ### Setting universal defaults
 
-You can set a new default for any option by setting `default_tweet_button_options` or `default_libe_button_options` or `default_pinit_button_options` in your application helper.For example:
+You can set a new default for any option in your `social_buttons.rb` initializer (in `config/initializers`.
+
+Set the `default_options` on any of the social view helper modules.
+
+For example:
 
 ```ruby
-    module ApplicationHelper
-      include TweetButton
-      TweetButton.default_tweet_button_options = {:via => "myself"}
-    end
+# config/initializers/social_buttons.rb
+SocialButtons::Tweet.default_options = {:via => "myself"}
 ```
-Only the options you specify will be overridden; so if you only specify a new default `:via`, then the other defaults will stay intact.
 
+Or even sweeter...
+
+```ruby
+# config/initializers/social_buttons.rb
+SocialButtons.config :tweet do |tweet|
+  tweet.default_options = {:via => "myself"}
+end
+
+# or
+SocialButtons.config(:tweet).default_options = {:via => "myself"}
+```
+
+Or simply :)
+
+```ruby
+# config/initializers/social_buttons.rb
+SocialButtons.config do |social|
+  social.tweet do |tweet|    
+    tweet.default_options = {:via => "myself"}
+  end
+
+  social.like.default_options = {:via => "myself"}
+  end
+end
+```
+
+This little DSL makes the configuration more extensible for the future ;)
+
+
+Only the options you specify will be overridden; so if you only specify a new default `:via`, then the other defaults will stay intact.
 
 # TODO
 
-* Convert this gem into railtie
 * Add tests :P
 * Give more control to customize buttons
 * Add more social buttons like g+.
-
 
 # Note on Patches/Pull Requests
