+++ lib/social_buttons/engine.rb
@@ -1,0 +1,11 @@
+module SocialButtons
+  module Rails
+    class Engine < ::Rails::Engine
+      intializer 'social buttons config' do 
+        SocialButtons.buttons.each do |name|
+          ActionView::Base.send :include, SocialButtons::ViewHelper
+        end        
+      end
+    end
+  end
+end
