+++ lib/social_buttons/view_helper.rb
@@ -1,0 +1,7 @@
+module SocialButtons
+  module ViewHelper
+    SocialButtons.names.each do |name|
+      self.send :include, "SocialButtons::#{name.to_s.camelize}Button"
+    end        
+  end
+end
